+++ packages/bricks/src/Field.tsx
@@ -10,7 +10,7 @@ import {
 } from "@ariakit/react/collection";
 import { Role } from "@ariakit/react/role";
 import { useStoreState } from "@ariakit/react/store";
-import { forwardRef } from "@stratakit/foundations/secret-internals";
+import { forwardRef } from "@stratakit/internal-utils/react";
 import cx from "classnames";
 import { useInit } from "./~utils.useInit.js";
 import Description from "./Description.js";
@@ -18,7 +18,7 @@ import { FieldCollection, FieldControlTypeContext } from "./Field.internal.js";
 import Label from "./Label.js";
 
 import type { CollectionItemProps } from "@ariakit/react/collection";
-import type { BaseProps } from "@stratakit/foundations/secret-internals";
+import type { BaseProps } from "@stratakit/internal-utils/props";
 import type {
 	CollectionStoreItem,
 	FieldCollectionStoreItem,