+++ packages/structures/src/Popover.tsx
@@ -9,17 +9,18 @@ import { PortalContext } from "@ariakit/react/portal";
 import { useStoreState } from "@ariakit/react/store";
 import { Button } from "@stratakit/bricks";
 import {
-	forwardRef,
+	useCloseWatcher,
 	usePopoverApi,
-	useUnreactiveCallback,
-} from "@stratakit/foundations/secret-internals";
+	useStableCallback,
+} from "@stratakit/internal-utils/hooks";
+import { forwardRef } from "@stratakit/internal-utils/react";
 import cx from "classnames";
 import { useInit } from "./~utils.useInit.js";
 
 import type {
 	BaseProps,
 	FocusableProps,
-} from "@stratakit/foundations/secret-internals";
+} from "@stratakit/internal-utils/props";
 
 // ----------------------------------------------------------------------------
 
@@ -78,10 +79,13 @@ const PopoverRoot = forwardRef<"div", PopoverRootProps>(
 		const store = AkPopover.usePopoverContext();
 		const popoverElement = useStoreState(store, "popoverElement");
 		const open = useStoreState(store, "open");
-		const setOpen = useUnreactiveCallback(store?.setOpen);
+		const setOpen = useStableCallback(store?.setOpen);
 
 		const popoverProps = usePopoverApi({
 			element: popoverElement,
+			open,
+		});
+		useCloseWatcher({
 			open,
 			setOpen,
 		});